refactoring: HeartHubPageViewController객체를 앱에서 전역적으로  재사용하기 위해서 구현
