+++ HeartHub/Sources/CommunityScene/Util/MockData.swift
@@ -16,20 +16,20 @@ struct MockData {
 
 let mockData = [
     MockData(
-        images: [UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!],
-        profileImage: UIImage(named: "test1")!,
+        images: [UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!],
+        profileImage: UIImage(named: "HeartBrand")!,
         profileName: "안녕",
         postLabel: "ㅇ아아아ㅏ안매아ㅔㅁ나애ㅔ마네애ㅏㄴㅁ"
     ),
     MockData(
-        images: [UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!],
-        profileImage: UIImage(named: "crown")!,
+        images: [UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!],
+        profileImage: UIImage(named: "HeartBrand")!,
         profileName: "허허",
         postLabel: "ㅇ아아아ㅏ안매아ㅔㅁ나애ㅔ마네애ㅏㄴㅁ"
     ),
@@ -40,28 +40,28 @@ let mockData = [
         postLabel: "ㅇ아아아ㅏ안매아ㅔㅁ나애ㅔ마네애ㅏㄴㅁ"
     ),
     MockData(
-        images: [UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!],
-        profileImage: UIImage(named: "test1")!,
+        images: [UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!],
+        profileImage: UIImage(named: "HeartBrand")!,
         profileName: "안녕",
         postLabel: "ㅇ아아아ㅏ안매아ㅔㅁ나애ㅔ마네애ㅏㄴㅁ"
     ),
     MockData(
-        images: [UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!],
-        profileImage: UIImage(named: "test1")!,
+        images: [UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!],
+        profileImage: UIImage(named: "HeartBrand")!,
         profileName: "안녕",
         postLabel: "ㅇ아아아ㅏ안매아ㅔㅁ나애ㅔ마네애ㅏㄴㅁ"
     ),
     MockData(
-        images: [UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!,
-                 UIImage(named: "test1")!],
+        images: [UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!,
+                 UIImage(named: "HeartBrand")!],
         profileImage: UIImage(systemName: "plus")!,
         profileName: "안녕",
         postLabel: "ㅇ아아아ㅏ안매아ㅔㅁ나애ㅔ마네애ㅏㄴㅁ"
